Update iOS CI build configuration in main.yml
Updated the `runs-on` parameter for the `build-ios` job to use `macos-26` instead of `macos-15`. Renamed the job from `mauibeach-android-ci-build` to `iOS Build` to reflect its purpose. Adjusted the `path` parameter to focus on iOS build artifacts. Added a `Checkout` step using `actions/checkout@v6` to ensure the repository is properly checked out during the build process.